pyproject.toml
[tool.poetry]
name = "notifications"
version = "0.1.0"
description = ""
authors = ["Dan <babindj@uriit.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
alembic = "^1.9.1"
sqlalchemy = "^1.4.46"
jinja2 = "^3.1.2"
dramatiq = {extras = ["rabbitmq", "watch"], version = "^1.13.0"}
pyproject-flake8 = "^6.0.0.post1"
aiohttp = "^3.8.3"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.5"
python-dateutil = "^2.8.2"
aiomisc = "^16.2.10"
jinja2schema = "^0.1.4"
cachetools = "^5.2.0"
sentry-sdk = "^1.12.1"
sentry-dramatiq = "^0.3.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
testcontainers = {extras = ["postgres"], version = "^3.7.1"}
isort = "^5.11.4"
black = "^22.12.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
httpx = "^0.23.3"

[tool.flake8]
ignore = ["E231", "E241", "E712", "E731", "F403"]
per-file-ignores = [
    "__init__.py:F401",
]
exclude = [
    "migrations/"
]
max-line-length = 120
count = true

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
